tools: introduce the prowjob analyzer
This new tool aim to assist QE and devs on making analysis of Prow CI
jobs of the project. It comprises of a set of scripts to fetch, parse
and generate data from a given prow job, and a command for Claude that
is going to use these scripts to effectively run analysis.

The scripts and command were generated by Claude itself. We feeded it
with some context, it generated PROWJOB_ANALYZER_PLAN.md, we approved
it and the magic happened. Afterwards we asked claude to fix and do some
adjustments here and there.

Look at scripts/prowjob-analyzer/README.md for further details.

Signed-off-by: Tom Buskey <tbuskey@redhat.com>
Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>